Traktor: Some more refactorizations, trying to keep IWorldRenderPass as slim as possible.
[git-p4: depot-paths = "//depot/traktor/": change = 21564]

+++ code/Mesh/MeshClassFactory.cpp
@@ -71,7 +71,7 @@ public:
 
 	void setVectorArrayParameter(const std::wstring& name, const AlignedVector< Vector4 >& param)
 	{
-		m_programParameters->setVectorArrayParameter(name, param.c_ptr(), param.size());
+		m_programParameters->setVectorArrayParameter(name, param.c_ptr(), (int)param.size());
 	}
 
 	void setTextureParameter(const std::wstring& name, render::ITexture* texture)
